% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{get_indicatrix.tissot}
\alias{get_indicatrix.tissot}
\title{This function returns the Tissot indicatrix circles that were generated}
\usage{
get_indicatrix.tissot(obj)
}
\arguments{
\item{obj}{tissot object}
}
\value{
Tissot indicatrix as polygon objects of class sf
}
\description{
This function returns the Tissot indicatrix circles that were generated
}
