% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{plot.tissot}
\alias{plot.tissot}
\title{This function plots the sf geometry and Tissot indicatrix circles that were generated}
\usage{
\method{plot}{tissot}(obj, crs = "auto")
}
\arguments{
\item{obj}{tissot object}

\item{crs}{target coordinate reference system: object of class 'crs', or input string for st_crs,
default is automatic}
}
\description{
This function plots the sf geometry and Tissot indicatrix circles that were generated
}
