% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{print.tissot}
\alias{print.tissot}
\title{This function prints Tissot indicatrix circles that were generated}
\usage{
\method{print}{tissot}(obj)
}
\arguments{
\item{obj}{tissot object}
}
\description{
This function prints Tissot indicatrix circles that were generated
}
