% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{[.indicatrix_list}
\alias{[.indicatrix_list}
\title{Subset an indicatrix_list}
\usage{
\method{[}{indicatrix_list}(x, i)
}
\arguments{
\item{x}{an \code{indicatrix_list}}

\item{i}{indices}
}
\value{
An \code{indicatrix_list} with the selected elements
}
\description{
Subset an indicatrix_list
}
