% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{summarize.tissot}
\alias{summarize.tissot}
\title{This function summarizes Tissot indicatrix circles details}
\usage{
summarize.tissot(obj)
}
\arguments{
\item{obj}{tissot object}
}
\description{
This function summarizes Tissot indicatrix circles details
}
