% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{ti_ellipse}
\alias{ti_ellipse}
\title{Generate ellipse coordinates for an indicatrix}
\usage{
ti_ellipse(x, scale = 1e+05, n = 72, ...)
}
\arguments{
\item{x}{an indicatrix object}

\item{scale}{scaling factor for the ellipse size in projected units}

\item{n}{number of points on the ellipse}

\item{...}{ignored}
}
\value{
A two-column matrix of projected coordinates tracing the ellipse
}
\description{
Generate ellipse coordinates for an indicatrix
}
