% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot-package.R
\docType{package}
\name{tissot-package}
\alias{tissot-package}
\alias{_PACKAGE}
\title{tissot: The Tissot Indicatrix}
\description{
Create and plot the Tissot Indicatrix.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/tissot}
  \item Report bugs at \url{https://github.com/hypertidy/tissot/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Bill Huber \email{whuber@quantdec.com}
}

}
\keyword{internal}
