% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot-package.R
\docType{data}
\name{world}
\alias{world}
\title{world coastline}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 82403 rows and 2 columns.
}
\usage{
world
}
\description{
A modified matrix version of data from the maps package.
}
\details{
Basically longitudes have been smooshed to 'abs(lon) < 180'
}
\keyword{datasets}
