% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{indicatrix}
\alias{indicatrix}
\alias{plot.indicatrix}
\title{Indicatrix}
\usage{
indicatrix(x, scale = 1, ...)

\method{plot}{indicatrix}(x, asp = 1, xlab = "Easting", ylab = "Northing",
  add = FALSE, ..., col.base = rgb(0, 0, 0, 0.1), col.lambda = grey(0.75),
  col.phi = "#45A271", col.major = "#A782C3", col.minor = "#C87A8A",
  col.outline = "black")
}
\arguments{
\item{x}{object from \code{tissot}}

\item{scale}{scaling}

\item{...}{arguments \code{n}, \code{from} and \code{to} passed to \code{ti_ellipse} function}
}
\description{
Indicatrix
}
\details{
Reprocesses the output of \code{tissot} into convenient geometrical data.
}

