% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tissot.R
\name{indicatrix}
\alias{indicatrix}
\title{Indicatrix}
\usage{
indicatrix(x, scale = 1, ...)
}
\arguments{
\item{x}{object from \code{tissot}}

\item{scale}{scaling}

\item{...}{arguments passed to internal function}
}
\description{
Indicatrix
}
\details{
Reprocesses the output of \code{tissot} into convenient geometrical data.
}

