% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{plot.indicatrix}
\alias{plot.indicatrix}
\title{Plot an indicatrix}
\usage{
\method{plot}{indicatrix}(
  x,
  scale = 1e+05,
  n = 72,
  col = "#FF990055",
  border = "black",
  add = TRUE,
  ...
)
}
\arguments{
\item{x}{an indicatrix object}

\item{scale}{scaling factor for the ellipse size}

\item{n}{number of points on the ellipse}

\item{col}{fill colour for the ellipse}

\item{border}{border colour}

\item{add}{logical; add to existing plot?}

\item{...}{passed to \code{\link[=polygon]{polygon()}}}
}
\description{
Plot an indicatrix
}
