% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{ti_ellipse}
\alias{ti_ellipse}
\title{Ellipse}
\usage{
ti_ellipse(center, axes, scale = 1, n = 36, from = 0, to = 2 * pi)
}
\arguments{
\item{center}{center}

\item{axes}{axes}

\item{scale}{scale}

\item{n}{n}

\item{from}{from}

\item{to}{to}
}
\value{
matrix
}
\description{
Ellipse
}
