% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot-package.R
\docType{package}
\name{tissot-package}
\alias{tissot-package}
\title{tissot: Tissot Indicatrix}
\description{
Attempt to formalize the Tissot indicatrix calculations shared by Bill Huber on Stack Exchange (\url{https://gis.stackexchange.com/a/5075/482}). Attempt to translate and automate the computation for any projection with a given set of longitude latitude locations. The indicatrix characterizes map projection distortion by computing scale factors, angular deformation, and convergence at each point.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/tissot}
  \item Report bugs at \url{https://github.com/hypertidy/tissot/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Authors:
\itemize{
  \item Bill Huber (original algorithm and calculations)
}

}
\keyword{internal}
