% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot-package.R
\docType{package}
\name{tissot-package}
\alias{tissot-package}
\title{tissot: Tissot Indicatrix}
\description{
Compute and visualize the Tissot indicatrix for map projections. The indicatrix characterizes projection distortion by computing scale factors, angular deformation, areal distortion, and convergence at arbitrary points. Based on the calculations shared by Bill Huber on Stack Exchange (\url{https://gis.stackexchange.com/a/5075/482}). Uses GDAL (via 'gdalraster') for coordinate transformation and supports any projection that GDAL/PROJ can handle.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/tissot}
  \item Report bugs at \url{https://github.com/hypertidy/tissot/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Authors:
\itemize{
  \item Bill Huber (original algorithm and calculations)
}

}
\keyword{internal}
