% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{Tissot}
\usage{
tissot(
  lambda,
  phi = NULL,
  degrees = TRUE,
  A = 6378137,
  f.inv = 298.257223563,
  ...,
  proj.in,
  proj.out
)
}
\arguments{
\item{lambda}{longitude}

\item{phi}{latitude}

\item{degrees}{logical, work in degrees or radians}

\item{A}{ellipsoidal semi-major axis (meters)}

\item{f.inv}{the inverse flattening}

\item{...}{passed to internal function}

\item{proj.in}{projection of input}

\item{proj.out}{projection of context}
}
\value{
list with stuff as per below
}
\description{
Create the Tissot Indicatrix.
}
\details{
Compute properties of scale distortion and Tissot's indicatrix at location x = c(lambda, phi)
using prj as the projection.  A is the ellipsoidal semi-major axis (in meters) and f.inv is
the inverse flattening.  The projection must return a vector (x, y) when given a vector (lambda, phi).
(Not vectorized.)  Optional arguments ... are passed to prj.
Source: Snyder pp 20-26 (WGS 84 defaults for the ellipsoidal parameters).
All input and output angles are in degrees.
}
\examples{
x <- seq(-175, 175, by = 15)
y <- seq(-82.5, 82.5, by = 15)
xy <- expand.grid(x, y)
r <- tissot(xy,
            proj.in= "OGC:CRS84",
            proj.out= "+proj=robin")

j <- which.min(abs(135 - r$lon) + abs(54 - r$lat))
idx0 <- indicatrix0(r[j, ], scale=10^4, n=71)
op <- par(mfrow =   c(2, 1))
plot(idx0, add = FALSE)
idx <- indicatrix(r, scale=3e5, n=71)
plot(idx, add = FALSE)
tissot_abline(r$lon[j], r$lat[j])
par(op)


}
