% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{The Tissot Indicatrix}
\usage{
tissot(lambda, phi, prj = function(z) z + 0, asDegrees = TRUE,
  A = 6378137, f.inv = 298.257223563, ...)
}
\arguments{
\item{lambda}{longitude}

\item{phi}{latitude}

\item{prj}{projection in PROJ.4}

\item{asDegrees}{logical, work in degrees or radians}

\item{A}{ellipsoidal semi-major axis (meters)}

\item{f.inv}{the inverse flattening}

\item{...}{passed to internal function}
}
\value{
list with stuff as per below
}
\description{
Create the Tissot Indicatrix.
}
\details{
Compute properties of scale distortion and Tissot's indicatrix at location x = c(lambda, phi)
using prj as the projection.  A is the ellipsoidal semi-major axis (in meters) and f.inv is
the inverse flattening.  The projection must return a vector (x, y) when given a vector (lambda, phi).
(Not vectorized.)  Optional arguments ... are passed to prj.
Source: Snyder pp 20-26 (WGS 84 defaults for the ellipsoidal parameters).
All input and output angles are in degrees.
}

