% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{Tissot}
\usage{
tissot(
  x,
  y = NULL,
  degrees = TRUE,
  A = 6378137,
  f.inv = 298.257223563,
  ...,
  proj.in,
  proj.out
)
}
\arguments{
\item{A}{ellipsoidal semi-major axis (meters)}

\item{f.inv}{the inverse flattening}

\item{...}{passed to internal function}

\item{proj.in}{projection of input}

\item{proj.out}{projection of context}

\item{lambda}{longitude}

\item{phi}{latitude}

\item{asDegrees}{logical, work in degrees or radians}
}
\value{
list with stuff as per below
}
\description{
Create the Tissot Indicatrix.
}
\details{
Compute properties of scale distortion and Tissot's indicatrix at location x = c(lambda, phi)
using prj as the projection.  A is the ellipsoidal semi-major axis (in meters) and f.inv is
the inverse flattening.  The projection must return a vector (x, y) when given a vector (lambda, phi).
(Not vectorized.)  Optional arguments ... are passed to prj.
Source: Snyder pp 20-26 (WGS 84 defaults for the ellipsoidal parameters).
All input and output angles are in degrees.
}
\examples{
x <- seq(-165, 165, by = 15)
y <- seq(-82.5, 82.5, by = 15)
xy <- expand.grid(x, y)
r <- tissot(xy,
            proj.in= "OGC:CRS84",
            proj.out= "+proj=robin")

i <- indicatrix0(r[1, ], scale=10^4, n=71)
plot(i, add = F)
ii <- indicatrix(r, scale=3e5, n=71)
plot(ii, add = F)



}
