% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{The Tissot Indicatrix}
\usage{
tissot(lambda, phi, prj = function(z) z + 0, asDegrees = TRUE,
  A = 6378137, f.inv = 298.257223563, ...)
}
\arguments{
\item{lambda}{longitude}

\item{phi}{latitude}

\item{prj}{projection in PROJ.4}

\item{asDegrees}{logical, work in degrees or radians}

\item{A}{ellipsoidal semi-major axis (meters)}

\item{f.inv}{the inverse flattening}

\item{...}{passed to internal function}
}
\value{
list with stuff as per below
}
\description{
Create the Tissot Indicatrix.
}
\details{
Compute properties of scale distortion and Tissot's indicatrix at location x = c(lambda, phi)
using prj as the projection.  A is the ellipsoidal semi-major axis (in meters) and f.inv is
the inverse flattening.  The projection must return a vector (x, y) when given a vector (lambda, phi).
(Not vectorized.)  Optional arguments ... are passed to prj.
Source: Snyder pp 20-26 (WGS 84 defaults for the ellipsoidal parameters).
All input and output angles are in degrees.
}
\examples{
library(rgdal)
prj <- function(z, proj.in, proj.out) {
  z.pt <- SpatialPoints(coords = matrix(z, ncol = 2), proj4string = proj.in)
  w.pt <- spTransform(z.pt, CRS = proj.out)
  coordinates(w.pt)[1,]
}
# Longitude, latitude, and reprojection function
# NAD 27 in
# World Robinson projection out
r <- tissot(130, 54, prj,
            proj.in=CRS("+init=epsg:4267"),
            proj.out=CRS("+init=esri:54030"))

i <- indicatrix(r, scale=10^4, n=71)
plot(i$outline, type="n", asp=1, xlab="Easting", ylab="Northing")
polygon(i$base, col=rgb(0, 0, 0, .025), border="Gray")
lines(i$d.lambda, lwd=2, col="Gray", lty=2)
lines(i$d.phi, lwd=2, col=rgb(.25, .7, .25), lty=2)
lines(i$axis.major, lwd=2, col=rgb(.25, .25, .7))
lines(i$axis.minor, lwd=2, col=rgb(.7, .25, .25))
lines(i$outline, asp=1, lwd=2)
}

