% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{Tissot}
\usage{
tissot(radius)
}
\arguments{
\item{radius}{double}
}
\value{
list
}
\description{
creates a geometry with equally sized circles (e.g. with 400 km radius) located on a regular grid in longitude / latitude
}
\examples{
tissot(2)
}
