% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{This function generates Tissot's indicatrix}
\usage{
tissot()
}
\value{
circles as objects of class sf
}
\description{
This function returns equally-sized circles (radius = 300km) as sf-polygons
to use as Tissot's indicatrix. The circles are located on a regular grid in
lon/lat every 20 degrees.
}
