% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot_map.R
\name{tissot_map}
\alias{tissot_map}
\alias{tissot_get_proj}
\title{Get last plot projection}
\usage{
tissot_map(..., add = TRUE)

tissot_get_proj()
}
\arguments{
\item{...}{graphical parameters for [lines()] if 'add = TRUE', or for [plot()] if 'add = FALSE'}

\item{add}{logical, default 'TRUE' add to existing plot or create new}
}
\value{
'tissot_map()' returns the internal world map data (projected if one is current) as a matrix

'tissot_get_proj()' returns the value of the current projection, or NULL
}
\description{
'tissot_map()' will add the [world] coastline to any map.
}
\details{
'tissot_get_proj()' When the indicatrix is plotted it registers its projection. This string
can be obtained with this getter function.
}
